Scheduler consumer as default implementation
