+++ prototype/src/main/java/sd/nosql/prototype/service/impl/QueueServiceImpl.java
@@ -1,16 +1,11 @@
 package sd.nosql.prototype.service.impl;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import sd.nosql.prototype.Record;
 import sd.nosql.prototype.enums.Operation;
 import sd.nosql.prototype.request.QueueRequest;
 import sd.nosql.prototype.service.PersistenceService;
 import sd.nosql.prototype.service.QueueService;
 
-import java.util.Date;
-import java.util.Timer;
-import java.util.TimerTask;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.Semaphore;
@@ -18,7 +13,6 @@ import java.util.concurrent.TimeUnit;
 
 public class QueueServiceImpl implements QueueService {
     private static final int TIMEOUT = 10;
-    private static final Logger logger = LoggerFactory.getLogger(DatabaseServiceImpl.class);
     private Semaphore firstSemaphore = new Semaphore(1, true);
     private Semaphore secondSemaphore = new Semaphore(1, true);
     private PersistenceService persistenceService = new FilePersistenceServiceImpl();
@@ -47,24 +41,6 @@ public class QueueServiceImpl implements QueueService {
                 leaveSecondCriticalZone();
             }
         }
-    }
-
-    @Override
-    public void scheduleConsumer(int persistenceTimeInMs) {
-        Timer timer = new Timer();
-        TimerTask timerTask = new TimerTask() {
-            @Override
-            public void run() {
-                try {
-                    logger.info("Consuming queue requests...");
-                    consumeAll();
-                }
-                catch (Exception e) {
-                    logger.info("Error on consuming queue requests", e);
-                }
-            }
-        };
-        timer.schedule(timerTask, new Date(), persistenceTimeInMs);
     }
 
     private boolean enterFirstCriticalZone() throws InterruptedException {
